+++ client/waatea2_client/lib/models/trainingattendance_model.dart
@@ -6,6 +6,8 @@ class TrainingAttendanceModel {
   final String club;
   final int attendanceCount;
   final bool current;
+  final String remarks;
+  final String review;
 
   TrainingAttendanceModel(
       {required this.pk,
@@ -14,16 +16,21 @@ class TrainingAttendanceModel {
       required this.date,
       required this.club,
       required this.attendanceCount,
-      required this.current});
+      required this.current,
+      required this.remarks,
+      required this.review});
 
   factory TrainingAttendanceModel.fromJson(Map<String, dynamic> json) {
     return TrainingAttendanceModel(
-        pk: json['pk'],
-        dayofyear: json['dayofyear'],
-        season: json['season'],
-        date: json['date'],
-        club: json['club'],
-        attendanceCount: json['attendanceCount'],
-        current: json['current']);
+      pk: json['pk'],
+      dayofyear: json['dayofyear'],
+      season: json['season'],
+      date: json['date'],
+      club: json['club'],
+      attendanceCount: json['attendanceCount'],
+      current: json['current'],
+      remarks: json['remarks'],
+      review: json['review'],
+    );
   }
 }